Josh/refactor utils (#1811)
* Move api tests to api directory

* EcrSummary directory

* Change my mind about ecrsummary

* Move services directory out of components

* Consolidate tests into a single directory

* ecr metadata service

* Consistent service naming

* Missed a ecrSummaryService

* Move evaluateEncounterDate to ecrSummaryService

* Move DataTableDisplay to ClinicalInfo component

* Remove DataTableDisplay from utils

* Remove stray comment

* Rename EcrMetadataService.ts to ecrMetadataService.ts

* Rename EcrSummaryService.ts to ecrSummaryService.ts

* Move services into same directory

* Delete containers/ecr-viewer/src/app/view-data/services directory

* Fix labservice import

* Fix labservice import

---------
