+++ containers/ecr-viewer/src/app/services/formatService.tsx
