+++ containers/ecr-viewer/src/app/tests/components/ExpandCollapseButtons.test.tsx
