+++ containers/ecr-viewer/src/app/tests/components/SideNav.test.tsx
