+++ containers/ecr-viewer/src/app/view-data/components/AccordionContainer.tsx
@@ -4,10 +4,8 @@ import {
   evaluateProviderData,
   evaluateClinicalData,
   evaluateDemographicsData,
-  evaluateEcrMetadata,
   PathMappings,
 } from "../../utils";
-import { evaluateLabInfoData } from "@/app/labs/utils";
 import Demographics from "./Demographics";
 import SocialHistory from "./SocialHistory";
 import UnavailableInfo from "./UnavailableInfo";
@@ -18,7 +16,9 @@ import { Bundle } from "fhir/r4";
 import React, { ReactNode } from "react";
 import { Accordion } from "@trussworks/react-uswds";
 import LabInfo from "@/app/view-data/components/LabInfo";
-import { formatString } from "@/app/format-service";
+import { formatString } from "@/app/services/formatService";
+import { evaluateEcrMetadata } from "../../services/ecrMetadataService";
+import { evaluateLabInfoData } from "@/app/services/labsService";
 
 type AccordionContainerProps = {
   children?: ReactNode;
